requirements_xauusd.txt
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
flask-sock>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
pytz>=2024.1
openpyxl>=3.1.0
python-dotenv>=1.0.0
